+++ kdevplatform/language/duchain/types/abstracttype.cpp
@@ -130,7 +130,7 @@ bool AbstractType::contains(const AbstractType* type) const
     return equals(type);
 }
 
-uint AbstractType::hash() const
+size_t AbstractType::hash() const
 {
     return KDevHash() << d_func()->typeClassId << d_func()->m_modifiers << d_func()->m_sizeOf
                       << d_func()->m_alignOfExponent;
