+++ kdevplatform/language/duchain/uses.cpp
@@ -34,7 +34,7 @@ public:
 
     UsesItem& operator=(const UsesItem& rhs) = delete;
 
-    unsigned int hash() const
+    size_t hash() const
     {
         //We only compare the declaration. This allows us implementing a map, although the item-repository
         //originally represents a set.
@@ -69,7 +69,7 @@ public:
         AverageSize = 30 //This should be the approximate average size of an Item
     };
 
-    unsigned int hash() const
+    size_t hash() const
     {
         return m_item.hash();
     }
